requirements.txt
pandas
PyYAML
jsonschema
pytest
